+++ app/src/main/java/nextstep/github/MainActivity.kt
@@ -3,7 +3,7 @@ package nextstep.github
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import nextstep.github.ui.screen.GithubRepositoryRoute
+import nextstep.github.ui.screen.repo.GithubRepositoryRoute
 import nextstep.github.ui.theme.GithubTheme
 
 class MainActivity : ComponentActivity() {
